backend/requirements.txt
langchain
langchain-community
langchain-google-genai
chromadb
sentence-transformers
streamlit
pypdf
python-docx
rank_bm25
python-dotenv
fastapi
uvicorn
python-multipart
requests
